backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic>=2.7.0,<3.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiosqlite==0.19.0
aiomysql==0.2.0
pymysql==1.1.0
redis==5.0.1
celery==5.3.4
minio==7.2.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
yt-dlp==2023.11.16
openai==1.3.7
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
ffmpeg-python==0.2.0
pillow==10.1.0
python-dotenv==1.0.0
pydantic-settings>=2.5.0
aiofiles>=23.2.0
email-validator>=2.0.0
# Audio processing dependencies
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
scipy==1.11.4
requests>=2.28.0
pydub==0.25.1
matplotlib==3.7.2
tqdm==4.66.1